src/anki_deck_generator/ingest/pdf.py
from __future__ import annotations

from pathlib import Path


def extract_text_from_pdf(path: Path) -> str:
    import fitz  # PyMuPDF

    doc = fitz.open(path)
    try:
        parts: list[str] = []
        for page in doc:
            parts.append(page.get_text())
        return "\n".join(parts)
    finally:
        doc.close()
