pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "anki-deck-generator"
version = "0.1.0"
description = "Convert Chinese study notes (PDF/Markdown/DOCX) to Anki vocabulary CSV"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "boto3>=1.35",
    "langchain>=0.3",
    "langchain-aws>=1.4",
    "langchain-core>=0.3",
    "pydantic>=2",
    "pydantic-settings>=2",
    "pymupdf>=1.24",
    "python-docx>=1.1",
    "python-dotenv>=1.0",
]

[project.optional-dependencies]
dev = ["pytest>=8"]

[project.scripts]
anki-notes-pipeline = "anki_deck_generator.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["src/anki_deck_generator"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
